[Xamarin.Android.Build.Tasks] Add coded error for armeabi (#2180)
Fixes: https://github.com/xamarin/xamarin-android/issues/2173
Context: https://github.com/xamarin/xamarin-android/pull/2106

This commit checks for the presence of the now unsupported `armeabi`
architecture in the `$(AndroidSupportedAbis)` property by using a
conditional error in a new target.

Another option would be to to change the error to a warning and
*ignore* the ";armeabi;" substring in the property, but that seems a
little risky for cases where users might not notice the warning.
They would probably be surprised later when they discovered that
their `.apk` files did not contain `armeabi` support.

I also debated on whether it would be better to move this check into
the `<BuildApk/>` task.  I decided that since it is just a small test
to look for a substring in an MSBuild property, using a conditional
`<Error/>` task in a small new target seems fine.