[C++] Main migration + file renaming. WILL NOT BUILD - EXPECTED
This commit renames C files to C++ ones and implements most of the changes
however it will NOT build. This is unfortunate but necessary and the reason is
that the git move detection algorithm has a threshold on the number of changes
to a file in order to consider it a move instead of a deletion of one file and
addition of a completely new one. Two files in `libmonodroid` sources exceed
this threshold - `util.c` and `dylib-mono.c`. Their C++ migration would mark
them as new files and we'd lose not only history but also the ability to
merge/rebase on changes from master. Those two files will be modified in the
next commit.