[Mono.Android-Tests] Improve AggregateException logging
Context: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/1180/testReport/junit/(root)/AndroidClientHandlerTests/Mono_Android_Tests__Xamarin_Android_NetTests_AndroidClientHandlerTests_Sanity_Tls_1_2_Url_WithMonoClientHandlerFails___Release/

`AggregateException`, like `Exception`, can be ~anything, and
represent an *actual failure* that needs to be fixed...or a
(hopefully) transient network failure that will disappear the next
time the test is run.

Which brings us to
`AndroidClientHandlerTests.Sanity_Tls_1_2_Url_WithMonoClientHandlerFails()`:
this test has been failing more often of late, but *we don't know
why*.  Our *guess* is that it's a transient network failure, as
*sometimes* the test works, but even if we assume that's the case,
we're still getting these "failures".

How do we fix the failures?  By detecting a transient network issue,
and ignoring the test.

How do we detect a transient network issue?

¯\_(ツ)_/¯

Which is the problem: *presumably* there is an exception somewhere
within the `AggregateException` which we could use to differentiate
between an *actual* failure and a transient network failure, but
*currently* we don't log exception thrown.  Meaning we have nothing to
ponder about or check.

Update `AggregateException` use so that if we `Assert.Fail()`, we also
log the exception as part of the failure message.