[Xamarin.Android.Build.Tasks] set $(AndroidSdkDirectory) if blank
Context: https://github.com/xamarin/xamarin-android/pull/2204

In d6009efa, I fixed the issue if `$(AndroidSdkDirectory)` is set to a
non-existent directory.

However, downstream in `monodroid`, there is a test project with an MSBuild target such as:

    <Exec Command="&quot;$(JavaSdkDirectory)\bin\javac&quot; -etc" />

`/p:JavaSdkDirectory` is not passed in here at the command line or
imported, and so `/bin/javac` is attempted to be used--which won't
quite work...

This brings up an important point.

`$(AndroidSdkDirectory)`, `$(AndroidNdkDirectory)` and
`$(JavaSdkDirectory)` were public properties that were set if blank.
So we can't just remove them! There could be unknown projects, NuGet
packages, etc. relying on them.

So I think the fix here, is to set these properties if they are blank.
Internally in xamarin-android's MSBuild targets, we should always use
the private underscore-prefixed versions such as
`$(_AndroidSdkDirectory)`.

I suspect we will need this change for `monodroid:master` and
`monodroid:d16-0-p1` to build.