[monodroid] Convert path to utf8 on windows.
The `d_name` property on the directory structture is not
a `char*` on windows but a `wchar_t`. As a result it might
not behave as expected when running on the desktop.

we should use `utils.utf16_to_utf8` to convert the values
over to a format which `copy_file_to_internal_location` can
take.