Bump Mono to fix builds on the Linux PR builders
Linux PR builders run on Debian instead of Ubuntu and because of an omission in
the Mono SDKs makefiles the bots tried to use MXE instead of the mingw
environment provided by the distribution. The Mono bump includes a fix for this.

Additionally, remove `clang` as a dependency for Debian and friends. This is
because llvm as used by the newer Mono versions fails to build with clang +
libstdc++ headers (it apparently expects libc++ headers), so until there's
another way to fix it, this is the best solution.