[Xamarin.Android.Build.Tasks] Only update the CustomView if the document changed.
While investigating a Designer Build breakage in monodroid we
hit an issue where the `ConvertCustomView` task was updating
a file even if nothing had changed in it.

As a result this causes a problem because we end up in a loop
where `_GenerateJavaStubs` and `_UpdateAndroidResgen` and the
designer target all update the same files over and over.

So lets try to avoid that by only updating the custom view
if the file actaually changes. However lets also keep the
modified date for the file which was written by
`_UpdateAndroidResgen`. This will ensure we dont keep running
the same targets over and over.