[Xamarin.Android.Build.Tests] Fix BuildHelper Root
Some of the tests in monodroid are failing because
it was assumed that the test `Root` folder would be
in the project output directory.

However with the new `XAConfig` file changes in XA that
is no longer the case. As a result some of the unit tests
fail because it is looking for files in the wrong place.

This commit changes the `Root` property of the `ProjectBuilder`
to use the one from `XAConfig`.