Update SQLite to version 3.52.2
The full list of changes since our previous release 3.9.2 can be found here:

  https://sqlite.org/changes.html

Additionally, bump the base NDK API platform we use to build native code to 14.
This is in line with the API level Mono runtime is built with as well as with
the minimum supported API in newer NDKs. Making the change part of this commit
because it also affects the way SQLite is built.