[monodroid] Migrate to C++ (#2089)
The purpose of this conversion is to start using a language that is
stricter and yet more powerful than C.  Stricter in type safety, in
separation of private and public functions, in data encapsulation.
Taking advantage of the C++ *language* features (as opposed to its
standard library)  allows us to write better structured,
safer and less fragile code.  C++ also allows us to implement safer
memory management (not part of this commit, see below).

This commit implements the basic, skeletal work on straightforward
C to C++ conversion:

  * Introduces classes to encompass a number of APIs that can be
    grouped together

  * fixes a number of issues with the old C code (void pointer
    arithmetic, incorrect type casts, comparisons of integers with
    different signs or of different sizes)

  * introduces strict type checking in the Mono interface API (it
    used to use `void*` pointers which could lead to confusion and
    hard to detect runtime/build time errors, see below)

The Mono API (found in the `DylibMono` class) was error prone, as
mentioned above, because it allowed code similar to the following
fragment to build without any warning or error from the compilers:

	void my_code (MonoDomain *domain) { }
	// ...
	MonoDomain *domain = get_domain ();
	my_code (&domain);

`MonoDomain` (and other Mono types) were `typedef`ed to be `void`
which made the above call "valid" (as a `void**` ISA `void*`).
A developer working with code wouldn't immediately know what was the
issue because of the `typedef` hiding the actual `MonoDomain` type.
During test execution, code in `monodroid-glue.cc` would mistakenly
pass a pointer to `MonoDomain` pointer (a.k.a `void**`) to the Mono
runtime caused the runtime to block while attempting to obtain a lock
on the domain where the domain did not exist and, obviously, not
initialized, so and the call couldn't succeed.

Thus the idea to change the typedefs to `struct XY` which introduces
a unique type, making all mistakes of this kind (as well as invalid
casts) obviously wrong and caught by the compiler.

This commit does *not* implement much in the are of memory management
because such changes would make porting code from master to the
branch increasingly complex and error prone.  For the same reason
`monodroid-glue.cc` is not split up into separate files (which is the
eventual plan) but merely introduces a handful of classes in that file.
After the commit is merged into master both of those changes will be
implemented to further take advantage of C++ capabilities.