[Xamarin.Android.Build.Tasks] Backport XA0115 error but make it a warning. (#2249)
Commit b1d81460 introduced a new error XA0115.  This error is raised
if the user has `armeabi` in their `$(AndroidSupportedAbis)`.

We need to do a similar thing in the d15-9 release.  However it
should not be an error but a warning.  It will turn into an error for
the next release.

This commit also introduces a warning if the user is using a
`$(TargetFrameworkVersion)` less than 19.  These platforms have a
very low market share and will be slowly removed from our build
system going forward.  This warning is to make sure that users
get a heads up before their application starts failing to build.