Bump to xamarin/sqlite:3.25.2@4ea4c691 (#2260)
The full list of changes since our previous release which contained
SQLite 3.9.2 can be found here (it's "only" nearly 3 years of fixes):

	https://sqlite.org/changes.html

Additionally, bump the base NDK API platform we use to build native
code to API-14.  This is in line with the API level Mono runtime is
built with as well as with the minimum supported API in newer NDKs.
Making the change part of this commit because it also affects the way
SQLite is built.