[sqlite] Use '/' as path separator in cmake on Windows
Fixes: https://github.com/xamarin/xamarin-android/issues/2271

cmake mandates a uniform way of presenting filesystem paths across all the
operating systems it supports - it requires that path segments are separated
with the `/` character. This commit makes sure that Windows build of the library
uses the expected path separator character but also makes sure to escape the `\`
character on non-Windows systems, should it happen to be part of the SQLite's
source directory path.