[monodroid] Revert PR #2267: Refactoring: make monodroid-glue.cc smaller
Context: https://github.com/xamarin/xamarin-android/pull/2267

Rebase and Merging PR #2267, which resulted in commits 861ab530 and
186e07b1, [broke the master build][0]:

	error G71DE2284: conflicting declaration 'constexpr const char xamarin::android::internal::AndroidSystem::SYSTEM_LIB_PATH []'
	error G6A49BD09: 'xamarin::android::internal::AndroidSystem::SYSTEM_LIB_PATH' declared 'constexpr' outside its class
	error GAE4BF799: declaration of 'const char* xamarin::android::internal::AndroidSystem::SYSTEM_LIB_PATH' outside of class is not definition [-fpermissive]
	error G24A5756C: 'readdir_r' was not declared in this scope
	error GD678582F: 'symlink' was not declared in this scope

This is from the MXE build of `src/monodroid`, wherein MXE with gcc 5
is used, which is significantly older than the clang/etc. compilers
that are normally used to build `src/monodroid`.

(This wasn't caught earlier because PR #2267 didn't have the
**full-mono-integration-build** label, and thus didn't try using MXE
to build `src/monodroid` for Windows (32-bit and 64-bit).)

Revert PR #2267 so that we can properly diagnose things.

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/1201/