[Xamarin.Android.Build.Tasks] Properly link Resources w/ Aapt2 (#2268)
Fixes: https://github.com/xamarin/xamarin-android/issues/2205

Scenario:

 1. Create a new Android project

 2. Set `$(AndroidUseAapt2)`=True in the `.csproj`

 3. Install any NuGet package that depends on
    [Xamarin.GooglePlayServices.Basement][0]

 4. Configure `google-services.json` with a **Build action** of
    `@(GoogleServicesJson)`.

 5. Have any Resource on app (layout, string, animation), and add an
    empty `anim/simple_animation.xml`:

        <set></set>

 6. Add the following lines to `MainActivity.OnCreate()`:

        var animation = Resource.Animation.simple_animation;
        var gcm       = Resource.String.gcm_defaultSenderId;

Clean the project and Rebuild.  The Rebuild fails, because
`Resource.String.gcm_defaultSenderId` exists but
`Resource.Animation.simple_animation` does not.

The problem here is that on the first build the `compiled.flata` archive only contains

	Archive:  compiled.flata
	Length   Method    Size  Cmpr    Date    Time   CRC-32   Name
	--------  ------  ------- ---- ---------- ----- --------  ----
	     400  Stored      400   0% 01-01-1980 01:00 8ade13a7  values_goog_svcs_json.arsc.flat
	--------          -------  ---                            -------
	     400              400   0%                            1 file

which suggested that something is overwriting the `compiled.flata`
archive created by `_CompileResources`.

The actual problem is that the `_CompileAndroidLibraryResources`
target is processing the directory for the `@(GooglePlayServicesJson)`
items.  As a result it's creating a `compiled.flata` archive in the
same place as the main Resource directory `compiled.flata` archive is
located.  As a result the `_CompileResources` target never runs.

This was because we hard coded `$(path)\..\compiled.flata` as the
location for the archive, which was based on the assumption that
library resources would exist in `obj\$(Configuration)\lp\xx\jl\res`.
This would mean the archive would end up in
`obj\$(Configuration)\lp\xx\jl`.

However for `@(GooglePlayServicesJson)`, its directory is created in
`obj\$(Configuration)\xx`. As a result the archive it created in
`obj\$(Configuration)`, i.e. in the same place as the main one!
And because it exists, we don't re-create it.

The solution is to move all the archives into a dedicated folder.
We generate a hash based on the path of the resource directory to
create a unique archive name.  We then output all the items into
`obj\$(Configuration)\flata`.  The main archive is still called
`compiled.flata` and is output in the same place.

[0]: https://www.nuget.org/packages/Xamarin.GooglePlayServices.Basement/70.1501.0-preview2