[Tests] Make network tests a bit more resilient
We often see networ test "failures" caused by unreachability of one of the
servers those tests use. This creates false negatives since this is not really
the test failing but rather the environment not being correct and preventing the
test from performing its actual task.
This commit catches the `WebException` thrown in connection failed cases and
merely ignores the test instead of failing it. Full exception message and stack
trace is printed as part of the ignore message.