[tests] Make network tests a bit more resilient (#2279)
We often see network test "failures" caused by unreachability of one
of the remote servers those tests use.  This creates false negatives
since this is not really the test failing but rather the environment
not being correct and preventing the test from performing its actual
task.

This commit catches the `WebException` type thrown in connection
failed cases (`WebException.Status==WebExceptionStatus.ConnectFailure`)
and merely ignores the test instead of failing it.

The full exception message and stack trace is in the ignore message.