[tests] fix AndroidUpdateResourcesTest.Issue2205 on Windows
Context: http://build.devdiv.io/2098745

The `Issue2205` test is failing on Windows with:

    Error message
        System.ArgumentException : Illegal characters in path.
    Stack trace
        at System.IO.Path.CheckInvalidPathChars(String path, Boolean checkAdditional)
        at System.IO.Path.Combine(String path1, String path2)
        at Xamarin.ProjectTools.XamarinProject.UpdateProjectFiles(String directory, IEnumerable`1 projectFiles, Boolean doNotCleanup)
        at Xamarin.ProjectTools.XamarinProject.Populate(String directory, IEnumerable`1 projectFiles)
        at Xamarin.ProjectTools.ProjectBuilder.Save(XamarinProject project, Boolean doNotCleanupOnUpdate, Boolean saveProject)
        at Xamarin.ProjectTools.ProjectBuilder.Build(XamarinProject project, Boolean doNotCleanupOnUpdate, String[] parameters, Boolean saveProject, Dictionary`2 environmentVariables)
        at Xamarin.Android.Build.Tests.AndroidUpdateResourcesTest.Issue2205(Boolean useAapt2)

Looking at the test, the `GoogleServicesJson` build item was declaring
JSON string contents as a filename. Whoops!

I guess that MacOS was OK with a JSON filename, while Windows was
certainly not.

I changed the `BuildItem` to use the `TextContent` property, and used
a filename of `googleservices.json`. I also let VS format the method,
so it matches our code formatting standards.