[Tests] More resilient network tests, take 2
The previous
commit (https://github.com/xamarin/xamarin-android/commit/608d950a703ddb527258565c2a0705aa37f71789)
solved only part of the issue with network tests which fail on connection to a
test server. Not only one test was left without any code to catch failed
connection attempts, all the other tests failed to catch a connection issue in
situations when the Java backend tries to connect to the server only after the
managed code attempts to wait on or obtain the result of the request from the
asynchronous task running the request.

This commit *should* solve the remaining problems.