[mono-runtimes] NuGets and Buffers and Memory, oh my! (#2280)
Context: https://github.com/mono/mono/issues/10768

Sometimes types are available from multiple assemblies.  Consider
the `System.Span<T>` type: it is "provided"/available from:

 1. The Xamarin.Android's `mscorlib.dll`.
 2. The Xamarin.Android `System.Memory.dll` *facade assembly*
    (if available; see also dc125a6f)
 3. The [`System.Memory` NuGet package][0].

If a project is built and uses `Span<T>`, which assembly should be
preferred?

The current answer ("use `mscorlib.dll`") is not understandable to
@jonpryor's feeble mind -- he guesses that it's due in part to the
`System.Memory.nuspec` file, which special-cases `MonoAndroid1.0` --
but the related issue is one of "forward consistency": how do we keep
things from breaking in the future?

The proposed solution is to use nuget conflict resolution to find our
facades when they are available and fall back to package
implementation when there is no inbox version.

This is *supported* by updating the `MonoAndroid`-profile
`v1.0/RedistList/FrameworkList.xml` file so that certain assemblies
are prioritized vs. what is currently known about NuGet packages.
Specifically, the `System.Buffers.dll` and `System.Memory.dll`
assemblies are given versions of `4.0.99.0` within `FrameworkList.xml`,
which is higher than the corresponding NuGet package-provided
assembly versions of `4.0.1.0`.  (Note it's the *assembly* version,
*not* the NuGet package version, which is important.)  Higher values
are more important, so a value of `4.0.99.0` will ensure that the
Xamarin.Android-provided assemblies are used instead of the NuGet
package assemblies, until the NuGet package assembly version is
updated to e.g. 4.1.0.

[0]: https://www.nuget.org/packages/System.Memory/