[Xamarin.Android.Build.Tasks] Add CopyResource items to FileWrites (#2288)
Not quite sure how we ever got away with this.  As part of the build
process we extract from `@(EmbeddedResource)` a bunch of `.java`
files into the intermediate output directory.

These are used for a number of things like multidex and bootstrapping
mono.  For some reason we *never* added those to the `@(FileWrites)`
group.  So there is a good chance these files will get deleted by an
incremental clean.

This was picked up by a unit test in monodroid which deals with
fast deployment.