[monodroid] Release local refs earlier
Should fix https://github.com/xamarin/xamarin-android/issues/2257

We should release local ref earlier in these cases, because with
higher number of assemblies we can overflow the local reference table
soon enough, before returning from native code.

In case of apks it is not as much probable, but why not release it and
not take a chance.