[Xamarin.Android.Build.Tasks] fix _CopyIntermediateAssemblies outputs (#2308)
Context: https://github.com/xamarin/xamarin-android/issues/2247
Context: c2d36812ed3d10a86570b15a68df891ecfd8e060

The `_CopyIntermediateAssemblies` target has a bug for incremental
builds, reproduced by the following:

 1. Build a Xamarin.Android app project
 2. `touch` the project's assembly in `$(IntermediateOutputPath)`
 3. Build a second time. `_CopyIntermediateAssemblies` will run as
    expected.
 4. Build a third time. `_CopyIntermediateAssemblies` still runs!

When in this state, `_CopyIntermediateAssemblies` will always run.
This is bad because it triggers other expensive targets like
`_UpdateAndroidResgen` *every time*.

I believe I introduced this in c2d36812, which was when I saw
`_CopyIntermediateAssemblies` taking more time than it used to.
It was a good tradeoff though, since it prevented the
`_UpdateAndroidResgen` target from running too often.

d15-9 does not have c2d36812.

To fix this problem properly, I introduced a new
`_copyintermediate.stamp` file to be used as the `Outputs` of the
target.  In addition to fixing our incremental build here, there
should be performance gains in only verifying the timestamp of one
file in `Outputs`.

The `Inputs` of the `_CopyIntermediateAssemblies` were also
incorrect, as it was not using the "satellite" assembly files as an
input.

This does not fully finish #2247, as there are two other targets
listed I need to investigate further.  These are likely unrelated to
the changes in c2d36812.