[build] use <ProjectReference /> instead of <MSBuild />
We have something weird going when building Xamarin.Android itself:

    msbuild Xamarin.Android.sln
    msbuild Xamarin.Android.sln

You would assume that the second command would be a no-op, but it's
not...

`xa-prep-tasks.csproj` and `Xamarin.Android.Tools.BootstrapTasks`
seems to build *every* time. You can see the `CoreCompile` target
running, which is even worse on Windows, where the output assembly
file might be locked!

I found two cases where we are calling the `<MSBuild/>` MSBuild task,
where `<ProjectReference/>` would be more appropriate:

- `android-toolchain.targets` - had an `<MSBuild />` call to build
  `Xamarin.Android.Tools.BootstrapTasks`, but it *also* had a
  `<ProjectReference/>`. Removing it worked fine, and solved the
  problem.
- `Mono.Android.targets` - was manually building `api-xml-adjuster`.
  `<ProjectReference/>` would have worked fine here, but you need
  `DependsOnTargets="ResolveReferences"` for it to work.

There are still additional MSBuild targets in our build that are
running every time, but this is a decent start to fix some of them.