[Xamarin.Android.Build.Tasks] Fix Error in `CompileToDalvik`.
When using the InstantRun feature we are seeing the following
error.

	error MSB6003: The specified task executable "java" could not be run. '{path}obj/Debug/lp/2/classes.zip' does not exist. Please rebuild the project.

This is because `_CompileToDalvikLibraryJars` runs to compile
the various `.jar` files into `.dex` files for each library.
This helps to speed up incremental builds.

However in manny cases the `classes.zip` will NOT exist
in these `lp` directories and nor should it.
So rather than throwing an error lets just not include
the file if it does not exist.